Changed skills JS script
