+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@firebase/firestore": "^3.4.7",
+    "async": "^3.2.3",
     "firebase": "^9.6.10",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.1",
