+++ public/index.html
@@ -35,18 +35,10 @@
     <script async="async" src="scripts/showMenu.js"></script>
     <script async="async" src="scripts/changeTheme.js"></script>
     <script async="async" src="scripts/scroll.js"></script>
-<!--    <script async="async" src="scripts/showSkills.js"></script>-->
+
     <script async="async" src="scripts/qualificationList.js"></script>
     <script async="async" src="scripts/showServices.js"></script>
     <script async="async" src="scripts/portfolioSwiper.js"></script>
     <script src="scripts/swiper-bundle.min.js"></script>
-<!--    <script src="https://cdn.jsdelivr.net/npm/save-ukraine@0.18.0/dist/umd/main.js"></script>-->
-<!--    <script>-->
-<!--      Ukraine.save({-->
-<!--        ribbon: 'TOP_LEFT',-->
-<!--        hasShadow: true,-->
-<!--        countries: [],-->
-<!--      });-->
-<!--    </script>-->
   </body>
 </html>
